refactor: update `horasDiarias` to `Long` and improve overtime calculation
- Changed `horasDiarias` datatype from `Integer` to `Long` across entities, DTOs, and services for consistency and precision.
- Replaced `save` with `saveAndFlush` in `HoraExtraServiceWrite` to ensure immediate persistence.
- Improved overtime value calculations by updating helper methods like `formatHorasToInterval`.
- Removed unused `formatHorasExtra` method and cleaned up redundant logic.

+++ src/main/java/cv/inps/rh/assiduidade/application/dto/HoraExtraDTO.java
@@ -21,28 +21,28 @@ import java.util.UUID;
 @IgrpDTO
 public class HoraExtraDTO  {
 
-  
-  
+
+
   private Long id ;
-  
-  
+
+
   private UUID colaborador ;
-  
-  
+
+
   private LocalDate dataInicio ;
-  
-  
+
+
   private LocalDate dataFim ;
-  
-  
-  private Integer horasDiaria ;
-  
-  
+
+
+  private Long horasDiaria ;
+
+
   private Integer percentagemHora ;
-  
-  
+
+
   private Integer valorDiario ;
-  
+
   @Valid
   private AnexoReqDTO documento ;
